+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * refactor: `PagerItemInterface` methods now returns itself, instead of
   `PageInterface`
+* fix(`QueryBuilderAdapter`): throws an exception if the query does not have an
+  order by clause.
 
 ## 0.5.0